requirements.txt
pathway
python-dotenv
google-generativeai
litellm==1.17.3
streamlit
beautifulsoup4
unstructured[all-docs]
